XEngine2021-V7.24.0.1001 Build1105
其他更新
    协议头
        增加函数声明周期管理宏定义,删除了没有使用的宏定义
    公用头
        删除了不需要的公用类型定义
        合并了一些类型定义
    开源库
        更新FFMPEG库应用地址
客户端组件
    客户端模块
        UDP和TCP的SELECT增加了一个获取完整包的函数
    SSL客户端模块
        SSL现在支持设置是由系统直接收发数据还是由用户确定了
        SSL现在支持使用证书链证书
        SSL修改了错误信息名称定义
        SSL修改扩展函数内部实现方式
        SSL增加了Send和Recv两个函数,直接获取数据的内存地址,而不需要开辟新内存了
帮助组件
    组包模块
        CHUNKED回调函数扩展了一个参数,支持标准和扩展组包CHUNKED回调了
        扩展的标注组包,扩展组包,自定义组包修正对句柄参数没有正确判断导致的问题
        标准组包,扩展组包,自定义组包现在获取数据如果没有后续数据导出的长度会被设置为0了
        标准组包修正CHUNK包如果只有30个字节的末尾包无法正确处理的问题
        标准组包修正某些时候尾部协议头分段接受导致没有CHUNK结束的问题
        扩展组包现在支持CHUNKED包处理了
    授权验证模块
        本地验证获取过期时间在无限制版本的时候不会返回错误了
标准组件
    HTTP模块
        HTTP和HTTP2扩展包管理器修正对句柄参数没有正确判断导致的问题
        HTTP2修改内部流属性结构,现在能够正确处理优先级包了
        HTTP2现在输出的流信息包含优先级了
        HTTP2完善错误信息
        HTTP2支持处理PUSH_PROMISE和GOAWAY包了
        HTTP2现在支持打包PUSH,GOAWAY和PING包了
        HTTP2现在获取客户端消息函数可以导出消息类型了
    WEBSOCKET模块
        扩展包管理器现在能够正确导出错误信息了
        扩展包管理器修正对句柄参数没有正确判断导致的问题
        标准包函数将被声明过期,未来会在某个版本移除
核心组件
    安全模块
        SSL服务现在支持不加载ROOT证书了
        SSL服务现在支持加载证书链证书模式了
        SSL服务现在可以设置客户端是否加载证书
        SSL服务修正接受连接可能某些错误没有错误码导出的问题
        SSL服务支持使用其他套接字而不是SSL套接字来发送和接受数据了
        SSL服务修正异步套接字一直无法ACCEPT的问题
        SSL服务修改接受连接函数,现在获得客户端证书信息参数可以为NULL
        SSL服务增加了获取客户端对应套接字的函数
        SSL服务增加了扩展函数库,取消了标准函数导出,现在支持多服务管理了
        SSL服务增加了Send和Recv两个函数直接获取数据的内存地址,而不需要开辟新内存了

+++ XEngine_Include/XEngine_HelpComponents/HelpCompress_Define.h
@@ -23,8 +23,8 @@ typedef LPVOID XZIP;
 typedef struct tag_HelpCompress_ZLibInfo
 {
     BOOL bIsFile;                                                         //真为文件,假为目录
-    ULONG nCompressSize;                                                  //压缩后大小
-    ULONG nUnCompressSize;                                                //压缩前大小
+    DWORD nCompressSize;                                                  //压缩后大小
+    DWORD nUnCompressSize;                                                //压缩前大小
     CHAR tszFileName[MAX_PATH];                                           //文件名
     CHAR tszComment[MAX_PATH];                                            //文件注释
     CHAR tszChangeTime[128];                                              //修改时间
